Mark tests as ported to the TCK
- Move `CASE` tests to own file
- Move tests of spatial functions to own file
- Remove tests with bugged assertions
