+++ community/cypher/acceptance/src/test/scala/org/neo4j/internal/cypher/acceptance/CaseExpressionAcceptanceTest.scala
@@ -1,0 +1,51 @@
+/*
+ * Copyright (c) 2002-2016 "Neo Technology,"
+ * Network Engine for Objects in Lund AB [http://neotechnology.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Neo4j is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package org.neo4j.internal.cypher.acceptance
+
+import org.neo4j.cypher.{ExecutionEngineFunSuite, NewPlannerTestSupport}
+
+class CaseExpressionAcceptanceTest extends ExecutionEngineFunSuite with NewPlannerTestSupport {
+
+  test("case should handle mixed number types") {
+    val query =
+      """WITH 0.5 AS x
+        |WITH (CASE WHEN x < 1 THEN 1 ELSE 2.0 END) AS x
+        |RETURN x + 1
+      """.stripMargin
+
+    val result = executeScalarWithAllPlannersAndCompatibilityMode[Long](query)
+
+    result should equal(2)
+  }
+
+  test("case should handle mixed types") {
+    val query =
+      """WITH 0.5 AS x
+        |WITH (CASE WHEN x < 1 THEN "wow" ELSE true END) AS x
+        |RETURN x + "!"
+      """.stripMargin
+
+    val result = executeScalarWithAllPlannersAndCompatibilityMode[String](query)
+
+    result should equal("wow!")
+  }
+
+
+}
